backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
xgboost==2.0.3
lightgbm==4.3.0
scikit-learn==1.5.0
scipy==1.13.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.100
python-multipart==0.0.9
pydantic-settings==2.3.4
apscheduler==3.10.4
httpx==0.27.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
